+++ tiled/_tests/test_writing.py
@@ -24,7 +24,7 @@ from starlette.status import (
 )
 
 from ..catalog import in_memory
-from ..catalog.adapter import CatalogContainerAdapter
+from ..catalog.adapter import CatalogContainerAdapter, WouldDeleteData
 from ..client import Context, from_context, record_history
 from ..mimetypes import PARQUET_MIMETYPE
 from ..queries import Key
@@ -510,7 +510,7 @@ async def test_delete(tree):
             key="delete_me",
         )
         nodes_before_delete = (await tree.context.execute("SELECT * from nodes")).all()
-        assert len(nodes_before_delete) == 1
+        assert len(nodes_before_delete) == 1 + 1  # +1 for the root node
         data_sources_before_delete = (
             await tree.context.execute("SELECT * from data_sources")
         ).all()
@@ -528,10 +528,12 @@ async def test_delete(tree):
                 key="delete_me",
             )
 
-        client.delete("delete_me")
+        with pytest.raises(WouldDeleteData):
+            client.delete_contents("delete_me")  # Cannot easily delete internal data
+        client.delete_contents("delete_me", external_only=False)
 
         nodes_after_delete = (await tree.context.execute("SELECT * from nodes")).all()
-        assert len(nodes_after_delete) == 0
+        assert len(nodes_after_delete) == 0 + 1  # the root node should still exist
         data_sources_after_delete = (
             await tree.context.execute("SELECT * from data_sources")
         ).all()
@@ -559,20 +561,20 @@ async def test_delete_non_empty_node(tree):
         # Cannot delete non-empty nodes
         assert "a" in client
         with fail_with_status_code(HTTP_409_CONFLICT):
-            client.delete("a")
+            client.delete_contents("a")
         assert "b" in a
         with fail_with_status_code(HTTP_409_CONFLICT):
-            a.delete("b")
+            a.delete_contents("b")
         assert "c" in b
         with fail_with_status_code(HTTP_409_CONFLICT):
-            b.delete("c")
+            b.delete_contents("c")
         assert "d" in c
         assert not list(d)  # leaf is empty
         # Delete from the bottom up.
-        c.delete("d")
-        b.delete("c")
-        a.delete("b")
-        client.delete("a")
+        c.delete_contents("d")
+        b.delete_contents("c")
+        a.delete_contents("b")
+        client.delete_contents("a")
 
 
 @pytest.mark.asyncio
@@ -585,15 +587,15 @@ async def test_write_in_container(tree):
         df = pandas.DataFrame({"a": [1, 2, 3]})
         b = a.write_dataframe(df, key="b")
         b.read()
-        a.delete("b")
-        client.delete("a")
+        a.delete_contents("b", external_only=False)
+        client.delete_contents("a")
 
         a = client.create_container("a")
         arr = numpy.array([1, 2, 3])
         b = a.write_array(arr, key="b")
         b.read()
-        a.delete("b")
-        client.delete("a")
+        a.delete_contents("b", external_only=False)
+        client.delete_contents("a")
 
         a = client.create_container("a")
         coo = sparse.COO(
@@ -601,8 +603,8 @@ async def test_write_in_container(tree):
         )
         b = a.write_sparse(coords=coo.coords, data=coo.data, shape=coo.shape, key="b")
         b.read()
-        a.delete("b")
-        client.delete("a")
+        a.delete_contents("b", external_only=False)
+        client.delete_contents("a")
 
         a = client.create_container("a")
         array = awkward.Array(
@@ -614,8 +616,8 @@ async def test_write_in_container(tree):
         )
         b = a.write_awkward(array, key="b")
         b.read()
-        a.delete("b")
-        client.delete("a")
+        a.delete_contents("b", external_only=False)
+        client.delete_contents("a")
 
 
 @pytest.mark.asyncio