+++ tiled/catalog/orm.py
@@ -64,9 +64,14 @@ class Node(Timestamped, Base):
 
     # This id is internal, never exposed to the client.
     id = Column(Integer, primary_key=True, index=True, autoincrement=True)
+    parent = Column(
+        Integer,
+        ForeignKey("nodes.id", name="fk_nodes_parent", ondelete="CASCADE"),
+        nullable=True,
+        index=True,  # An index to support parent lookups
+    )
 
     key = Column(Unicode(1023), nullable=False)
-    ancestors = Column(JSONVariant, nullable=True)
     structure_family = Column(Enum(StructureFamily), nullable=False)
     metadata_ = Column("metadata", JSONVariant, nullable=False)
     specs = Column(JSONVariant, nullable=False)
@@ -81,17 +86,21 @@ class Node(Timestamped, Base):
     )
     revisions = relationship(
         "Revision",
-        backref="revisions",
+        backref="node",
         passive_deletes=True,
     )
 
+    # This is a self-referencing relationship between parent and children
+    prnt_rel = relationship("Node", remote_side=[id], back_populates="chld_rel")
+    chld_rel = relationship("Node", back_populates="prnt_rel")
+
     __table_args__ = (
-        UniqueConstraint("key", "ancestors", name="key_ancestors_unique_constraint"),
+        UniqueConstraint("key", "parent", name="key_parent_unique_constraint"),
         # This index supports comparison operations (==, <, ...).
         # For key-existence operations we will need a GIN index additionally.
         Index(
             "top_level_metadata",
-            "ancestors",
+            "parent",
             # include the keys of the default sorting ('time_created', 'id'),
             # used to avoid creating a temp sort index
             "time_created",
@@ -100,8 +109,28 @@ class Node(Timestamped, Base):
             "access_blob",
             postgresql_using="gin",
         ),
-        # This is used by ORDER BY with the default sorting.
-        # Index("ancestors_time_created", "ancestors", "time_created"),
+    )
+
+
+class NodesClosure(Base):
+    """This describes the closure table for Node."""
+
+    __tablename__ = "nodes_closure"
+
+    ancestor = Column(
+        Integer, ForeignKey("nodes.id", ondelete="CASCADE"), primary_key=True
+    )
+    descendant = Column(
+        Integer, ForeignKey("nodes.id", ondelete="CASCADE"), primary_key=True
+    )
+    depth = Column(Integer, nullable=False)
+
+    __table_args__ = (
+        UniqueConstraint(
+            "ancestor", "descendant", name="ancestor_descendant_unique_constraint"
+        ),
+        Index("idx_nodes_closure_ancestor", "ancestor"),
+        Index("idx_nodes_closure_descendant", "descendant"),
     )
 
 
@@ -274,6 +303,69 @@ def create_index_metadata_tsvector_search(target, connection, **kw):
                 """
             )
         )
+
+
+@event.listens_for(NodesClosure.__table__, "after_create")
+def update_closure_table(target, connection, **kw):
+    if connection.engine.dialect.name == "sqlite":
+        # Create a trigger to update the closure table when INSERTING a new node
+        connection.execute(
+            text(
+                """
+CREATE TRIGGER update_closure_table_when_inserting
+AFTER INSERT ON nodes
+BEGIN
+    INSERT INTO nodes_closure(ancestor, descendant, depth)
+    SELECT NEW.id, NEW.id, 0;
+    INSERT INTO nodes_closure(ancestor, descendant, depth)
+    SELECT p.ancestor, c.descendant, p.depth+c.depth+1
+    FROM nodes_closure p, nodes_closure c
+    WHERE p.descendant=NEW.parent and c.ancestor=NEW.id;
+END"""
+            )
+        )
+
+    elif connection.engine.dialect.name == "postgresql":
+        # Create function and trigger to update the closure table when INSERTING a new node
+        connection.execute(
+            text(
+                """
+CREATE OR REPLACE FUNCTION update_closure_table_when_inserting()
+RETURNS TRIGGER AS $$
+BEGIN
+    INSERT INTO nodes_closure(ancestor, descendant, depth)
+    VALUES (NEW.id, NEW.id, 0);
+    INSERT INTO nodes_closure(ancestor, descendant, depth)
+    SELECT p.ancestor, c.descendant, p.depth + c.depth + 1
+    FROM nodes_closure p, nodes_closure c
+    WHERE p.descendant = NEW.parent AND c.ancestor = NEW.id;
+    RETURN NEW;
+END;
+$$ LANGUAGE plpgsql;
+"""
+            )
+        )
+
+        connection.execute(
+            text(
+                """
+CREATE TRIGGER update_closure_table_when_inserting
+AFTER INSERT ON nodes
+FOR EACH ROW
+EXECUTE FUNCTION update_closure_table_when_inserting();
+"""
+            )
+        )
+
+    # Create the root node (in nodes and, automatically, in nodes_closure) when initializing the database.
+    connection.execute(
+        text(
+            """
+INSERT INTO nodes(id, key, parent, structure_family, metadata, specs, access_blob)
+SELECT 0, '', NULL, 'container', '{}', '[]', '{}';
+"""
+        )
+    )
 
 
 class FTS5Table(Table):