+++ tiled/server/router.py
@@ -1510,6 +1510,16 @@ def get_router(
     async def delete(
         request: Request,
         path: str,
+        recursive: Optional[bool] = Query(
+            False, description="Delete children recursively"
+        ),
+        external_only: Optional[bool] = Query(
+            True,
+            description=(
+                "Delete the node, but only if this would not "
+                "affect any internally-managed data sources"
+            ),
+        ),
         principal: Union[Principal, SpecialUsers] = Depends(get_current_principal),
         root_tree=Depends(get_root_tree),
         session_state: dict = Depends(get_session_state),
@@ -1528,7 +1538,7 @@ def get_router(
             getattr(request.app.state, "access_policy", None),
         )
         if hasattr(entry, "delete"):
-            await entry.delete()
+            await entry.delete(recursive=recursive, external_only=external_only)
         else:
             raise HTTPException(
                 status_code=HTTP_405_METHOD_NOT_ALLOWED,
@@ -1536,35 +1546,35 @@ def get_router(
             )
         return json_or_msgpack(request, None)
 
-    @router.delete("/nodes/{path:path}")
-    async def bulk_delete(
-        request: Request,
-        path: str,
-        principal: Union[Principal, SpecialUsers] = Depends(get_current_principal),
-        root_tree=Depends(get_root_tree),
-        session_state: dict = Depends(get_session_state),
-        authn_scopes: Scopes = Depends(get_current_scopes),
-        _=Security(check_scopes, scopes=["write:data", "write:metadata"]),
-    ):
-        entry = await get_entry(
-            path,
-            ["write:data", "write:metadata"],
-            principal,
-            authn_scopes,
-            root_tree,
-            session_state,
-            request.state.metrics,
-            None,
-            getattr(request.app.state, "access_policy", None),
-        )
-        if hasattr(entry, "delete_tree"):
-            await entry.delete_tree()
-        else:
-            raise HTTPException(
-                status_code=HTTP_405_METHOD_NOT_ALLOWED,
-                detail="This node does not support bulk deletion.",
-            )
-        return json_or_msgpack(request, None)
+    # @router.delete("/nodes/{path:path}")
+    # async def bulk_delete(
+    #     request: Request,
+    #     path: str,
+    #     principal: Union[Principal, SpecialUsers] = Depends(get_current_principal),
+    #     root_tree: pydantic_settings.BaseSettings = Depends(get_root_tree),
+    #     session_state: dict = Depends(get_session_state),
+    #     authn_scopes: Scopes = Depends(get_current_scopes),
+    #     _=Security(check_scopes, scopes=["write:data", "write:metadata"]),
+    # ):
+    #     entry = await get_entry(
+    #         path,
+    #         ["write:data", "write:metadata"],
+    #         principal,
+    #         authn_scopes,
+    #         root_tree,
+    #         session_state,
+    #         request.state.metrics,
+    #         None,
+    #         getattr(request.app.state, "access_policy", None),
+    #     )
+    #     if hasattr(entry, "delete_tree"):
+    #         await entry.delete_tree()
+    #     else:
+    #         raise HTTPException(
+    #             status_code=HTTP_405_METHOD_NOT_ALLOWED,
+    #             detail="This node does not support bulk deletion.",
+    #         )
+    #     return json_or_msgpack(request, None)
 
     @router.put("/array/full/{path:path}")
     async def put_array_full(
@@ -1945,7 +1955,7 @@ def get_router(
             drop_revision=drop_revision,
         )
 
-        response_data = {"id": entry.key}
+        response_data = {"id": entry.node.key}
         if metadata_modified:
             response_data["metadata"] = metadata
         if access_blob_modified:
@@ -2025,7 +2035,7 @@ def get_router(
             drop_revision=drop_revision,
         )
 
-        response_data = {"id": entry.key}
+        response_data = {"id": entry.node.key}
         if metadata_modified:
             response_data["metadata"] = metadata
         if access_blob_modified: