+++ CHANGELOG.md
@@ -4,6 +4,14 @@ Write the date in place of the "Unreleased" in the case a new version is release
 # Changelog
 
 
+## v0.1.0-b31 (Unreleased)
+
+### Changed
+
+- The logic of hierarchical organization of the Nodes table in Catalog: use the concept
+  of Closure Table to track ancestors and descendands of the nodes.
+
+
 ## v0.1.0-b31 (2025-08-01)
 
 ### Added
@@ -20,6 +28,7 @@ Write the date in place of the "Unreleased" in the case a new version is release
 - Typehinted database access methods
 - Explicit type conversion in SQL adapter when appending to an existing table.
 
+
 ## v0.1.0-b30 (2025-07-18)
 
 ### Changed
@@ -34,6 +43,7 @@ Write the date in place of the "Unreleased" in the case a new version is release
 - Restored authentication check for API key
 - Updated usage for change in Zarr 3.x API.
 - Improved error message if config location is non-file
+
 
 ## v0.1.0-b29 (2025-06-06)